R1/requirements.txt
# R1 Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.27
python-multipart>=0.0.6

# Browser Automation
playwright>=1.40

# Chat Integrations
websockets>=12.0
twilio>=8.10
discord.py>=2.3.0
slack-sdk>=3.21.0
python-telegram-bot>=20.0
aiohttp>=3.9.0

# System
psutil>=5.9
pytz>=2024.1
APScheduler>=3.10.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Config
pyyaml>=6.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.0
watchdog>=3.0.0
packaging>=24.0
numpy>=2.0.0
SpeechRecognition>=3.10.0
pyautogui>=0.9.54
pyperclip>=1.9.0

# CLI
click>=8.1.0
rich>=13.7.0

# Security
cryptography>=42.0.0

# OpenClaw Additional Dependencies
pyttsx3>=2.90  # Text-to-speech
gTTS>=2.5.0  # Google TTS fallback
whisper>=1.1.0  # OpenAI Whisper for STT
edge-tts>=6.1.0  # Edge TTS
beautifulsoup4>=4.12.0  # Web scraping
requests>=2.31.0  # HTTP requests
langdetect>=1.0.9  # Language detection
deep-translator>=1.11.0  # Translation
